Fixed Java migrations from different packages.
Fix #32654

(cherry picked from commit cc1d756f7cdc66ea49c84deef373bd62f1293bd7)
