+++ extensions/flyway/deployment/src/test/java/db/migration/V1_0_1__Update.java
@@ -1,0 +1,18 @@
+package db.migration;
+
+import java.sql.Statement;
+
+import org.flywaydb.core.api.migration.BaseJavaMigration;
+import org.flywaydb.core.api.migration.Context;
+
+/**
+ * Migration class for some testcases.
+ */
+public class V1_0_1__Update extends BaseJavaMigration {
+    @Override
+    public void migrate(Context context) throws Exception {
+        try (Statement statement = context.getConnection().createStatement()) {
+            statement.executeUpdate("INSERT INTO quarked_flyway VALUES (1001, 'test')");
+        }
+    }
+}
