+++ extensions/flyway/deployment/src/test/java/io/quarkus/flyway/test/FlywayExtensionCleanAndMigrateAtStartWithJavaMigrationTest.java
@@ -12,13 +12,14 @@ import jakarta.inject.Inject;
 
 import org.flywaydb.core.Flyway;
 import org.flywaydb.core.api.MigrationVersion;
-import org.flywaydb.core.api.migration.BaseJavaMigration;
 import org.flywaydb.core.api.migration.Context;
 import org.flywaydb.core.api.migration.JavaMigration;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
+import db.migration.V1_0_1__Update;
+import db.migration.V1_0_2__Update;
 import io.agroal.api.AgroalDataSource;
 import io.quarkus.test.QuarkusUnitTest;
 
@@ -33,7 +34,7 @@ public class FlywayExtensionCleanAndMigrateAtStartWithJavaMigrationTest {
     @RegisterExtension
     static final QuarkusUnitTest config = new QuarkusUnitTest()
             .withApplicationRoot((jar) -> jar
-                    .addClasses(V1_0_1__Update.class, V1_0_2__Update.class)
+                    .addClasses(V1_0_1__Update.class, V1_0_2__Update.class, V9_9_9__Update.class)
                     .addAsResource("db/migration/V1.0.0__Quarkus.sql")
                     .addAsResource("clean-and-migrate-at-start-config.properties", "application.properties"));
 
@@ -53,19 +54,10 @@ public class FlywayExtensionCleanAndMigrateAtStartWithJavaMigrationTest {
         assertEquals("1.0.2", currentVersion, "Expected to be 1.0.2 as there is a SQL and two Java migration scripts");
     }
 
-    public static class V1_0_1__Update extends BaseJavaMigration {
-        @Override
-        public void migrate(Context context) throws Exception {
-            try (Statement statement = context.getConnection().createStatement()) {
-                statement.executeUpdate("INSERT INTO quarked_flyway VALUES (1001, 'test')");
-            }
-        }
-    }
-
-    public static class V1_0_2__Update implements JavaMigration {
+    public static class V9_9_9__Update implements JavaMigration {
         @Override
         public MigrationVersion getVersion() {
-            return MigrationVersion.fromVersion("1.0.2");
+            return MigrationVersion.fromVersion("9.9.9");
         }
 
         @Override
@@ -86,7 +78,7 @@ public class FlywayExtensionCleanAndMigrateAtStartWithJavaMigrationTest {
         @Override
         public void migrate(Context context) throws Exception {
             try (Statement statement = context.getConnection().createStatement()) {
-                statement.executeUpdate("INSERT INTO quarked_flyway VALUES (1002, 'test')");
+                statement.executeUpdate("INSERT INTO quarked_flyway VALUES (9999, 'should-not-be-added')");
             }
         }
     }
