+++ extensions/flyway/deployment/src/test/resources/clean-and-migrate-at-start-with-fs-resource-config.properties
@@ -10,4 +10,4 @@ quarkus.flyway.table=test_flyway_history
 quarkus.flyway.baseline-on-migrate=false
 quarkus.flyway.baseline-version=0.0.1
 quarkus.flyway.baseline-description=Initial description for test
-quarkus.flyway.locations=filesystem:src/test/resources/db/migration
+quarkus.flyway.locations=filesystem:src/test/resources/db/migration,classpath:db/migration
